set up logging level.
