+++ python/graphstorm/run/gsgnn_ep/gsgnn_ep.py
@@ -17,6 +17,8 @@
 """
 
 import os
+import logging
+
 import graphstorm as gs
 from graphstorm.config import get_argument_parser
 from graphstorm.config import GSConfig
@@ -55,6 +57,7 @@ def main(config_args):
     config = GSConfig(config_args)
     config.verify_arguments(True)
 
+    logging.basicConfig(level=args.logging_level)
     gs.initialize(ip_config=config.ip_config, backend=config.backend)
     rt_profiler.init(config.profile_path, rank=gs.get_rank())
     sys_tracker.init(config.verbose, rank=gs.get_rank())
