+++ python/graphstorm/run/gsgnn_lp/lp_infer_lm.py
@@ -17,6 +17,8 @@
     encoder only.
 """
 
+import logging
+
 import graphstorm as gs
 from graphstorm.config import get_argument_parser
 from graphstorm.config import GSConfig
@@ -34,6 +36,8 @@ def main(config_args):
     """
     config = GSConfig(config_args)
     config.verify_arguments(False)
+
+    logging.basicConfig(level=args.logging_level)
     gs.initialize(ip_config=config.ip_config, backend=config.backend)
     device = setup_device(config.local_rank)
