+++ python/graphstorm/run/gsgnn_np/np_infer_gnn.py
@@ -16,6 +16,8 @@
     Inference script for node classification/regression tasks with GNN
 """
 
+import logging
+
 import graphstorm as gs
 from graphstorm.config import get_argument_parser
 from graphstorm.config import GSConfig
@@ -42,6 +44,8 @@ def main(config_args):
     """
     config = GSConfig(config_args)
     config.verify_arguments(False)
+
+    logging.basicConfig(level=args.logging_level)
     gs.initialize(ip_config=config.ip_config, backend=config.backend)
     device = setup_device(config.local_rank)
