+++ python/graphstorm/utils.py
@@ -216,6 +216,21 @@ def print_mem(device):
         logging.info("Peak RAM Mem alloc: %.4f MB",
             resource.getrusage(resource.RUSAGE_SELF).ru_maxrss / 1024)
 
+def get_log_level(log_level):
+    """ Map the logging level.
+    """
+    if log_level == "debug":
+        return logging.DEBUG
+    elif log_level == "info":
+        return logging.INFO
+    elif log_level == "warning":
+        return logging.WARNING
+    elif log_level == "error":
+        return logging.ERROR
+    else:
+        raise ValueError(f"Unknown logging level {log_level}. " + \
+                "The possible values are: debug, info, warning, error.")
+
 class SysTracker:
     """ This tracks the system performance.
